+++ resources/config.yml
@@ -14,6 +14,10 @@
 #German (deu)
 language: eng
 
+#Specify here how you want to save the data
+#Available: sqlite, mysql
+datamanager: sqlite
+
 #MySQL Connection
 host: 127.0.0.1
 username: root
